LPS-117536 Extract StructureTree component
