+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/plugins/page-structure/components/Plugin.scss
@@ -1,2 +1,3 @@
 @import './PageStructureSidebar';
+@import './StructureTree';
 @import './StructureTreeNode';
